Add PROXY protocol NOOP TLV helpers
