+++ actix-proxy-protocol/src/tlv.rs
@@ -134,41 +134,8 @@ impl Tlv for Crc32c {
     }
 }
 
-/// The TLV of this type should be ignored when parsed. The value is zero or more
-/// bytes. Can be used for data padding or alignment. Note that it can be used
-/// to align only by 3 or more bytes because a TLV can not be smaller than that.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct Noop {
-    value: Vec<u8>,
-}
-
-impl Noop {
-    ///
-    ///
-    /// # Panics
-    /// Panics if `value` is empty (i.e., has length of 0).
-    pub fn new(value: impl Into<Vec<u8>>) -> Self {
-        let value = value.into();
-
-        assert!(!value.is_empty(), "Noop TLV `value` cannot be empty");
-
-        Self { value }
-    }
-}
-
-impl Tlv for Noop {
-    const TYPE: u8 = PP2_TYPE_NOOP;
-
-    fn try_from_value(value: &[u8]) -> Option<Self> {
-        Some(Self {
-            value: value.to_owned(),
-        })
-    }
-
-    fn value_bytes(&self) -> Cow<'_, [u8]> {
-        Cow::Borrowed(&self.value)
-    }
-}
+mod noop;
+pub use self::noop::Noop;
 
 /// The value of the type PP2_TYPE_UNIQUE_ID is an opaque byte sequence of up to
 /// 128 bytes generated by the upstream proxy that uniquely identifies the
